python-flask-hana/requirements.txt
Flask==3.0.*
cfenv==0.5.3
hdbcli==2.17.*
sap-xssec==4.*